ocr-service/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Azure Form Recognizer
azure-ai-formrecognizer==3.3.0
azure-core==1.29.5

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Image processing
pillow>=10.0.0
opencv-python==4.8.1.78

# OCR libraries
pytesseract==0.3.10
pdfplumber==0.10.3
pymupdf==1.23.8
pdf2image==1.16.3

# Utilities
python-dateutil==2.8.2

